refactoring code architecture to Blueprints
