+++ app/admin/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+admin_bp = Blueprint('admin', __name__, template_folder='templates')
+
+from . import routes
