+++ data/comments_data.csv
@@ -9,3 +9,6 @@ Estou travado no problema de Binary Search Tree,harry1,oi,26/01/2026 18:51
 Finalmente entendi como funciona o algoritmo de Dijkstra,harry2,"eai 
 ",26/01/2026 20:05
 Estou travado no problema de Binary Search Tree,harry2,ioruewiou,26/01/2026 20:05
+Consegui resolver o Two Sum usando HashMap!,num2,brabo man,29/01/2026 22:01
+Finalmente entendi como funciona o algoritmo de Dijkstra,num2,c++ on top,29/01/2026 22:01
+p1,num1,comentario 1,29/01/2026 22:36
