+++ templates/main/landing.html
@@ -16,8 +16,8 @@
                 Conte o que está pegando no LeetCode, compartilhe vitórias pequenas e receba feedback de quem também está virando noites nos estudos.
             </p>
             <div class="hero-actions">
-                <a href="{{ url_for('home') }}" class="btn btn-lg btn-primary">Começar agora</a>
-                <a href="{{ url_for('home') }}" class="btn btn-lg btn-outline-secondary">Já faço parte</a>
+                <a href="{{ url_for('main.home') }}" class="btn btn-lg btn-primary">Começar agora</a>
+                <a href="{{ url_for('main.home') }}" class="btn btn-lg btn-outline-secondary">Já faço parte</a>
             </div>
             <div class="d-flex align-items-center gap-3">
                 <div class="avatar-list">
@@ -86,8 +86,8 @@
             <h3 class="fw-bold">Topa postar seu próximo código?</h3>
             <p class="text-muted mb-4">Conta em minutos, comunidade para sempre. Priorize o processo e não só o resultado.</p>
             <div class="d-flex justify-content-center gap-3 flex-wrap">
-                <a href="{{ url_for('home') }}" class="btn btn-primary btn-lg">Quero entrar</a>
-                <a href="{{ url_for('home') }}" class="btn btn-outline-secondary btn-lg">Espiar a comunidade</a>
+                <a href="{{ url_for('main.home') }}" class="btn btn-primary btn-lg">Quero entrar</a>
+                <a href="{{ url_for('main.home') }}" class="btn btn-outline-secondary btn-lg">Espiar a comunidade</a>
             </div>
         </div>
     </section>
