+++ templates/main/post.html
@@ -9,12 +9,12 @@
         <aside class="col-12 col-md-3 col-lg-2 home-sidebar px-2 border-bottom border-md-0">
             <ul class="nav flex-column p-3">
                 <li class="nav-item mb-2">
-                    <a class="nav-link" href="{{ url_for('home') }}">
+                    <a class="nav-link" href="{{ url_for('main.home') }}">
                         <i class="bi bi-house"></i> Home
                     </a>
                 </li>
                 <li class="nav-item mb-2">
-                    <a class="nav-link" href="{{ url_for('publish') }}">
+                    <a class="nav-link" href="{{ url_for('main.publish') }}">
                         <i class="bi bi-plus-circle"></i> Criar Post
                     </a>
                 </li>
@@ -34,7 +34,7 @@
         <!-- Conteudo  -->
         <section class="col-12 col-md-9 col-lg-10 home-main-area">
             <div class="feed-container">
-                <a href="{{ url_for('home') }}" class="btn btn-outline-secondary mb-3">
+                <a href="{{ url_for('main.home') }}" class="btn btn-outline-secondary mb-3">
                     <i class="bi bi-arrow-left"></i> Voltar
                 </a>
 
@@ -92,7 +92,7 @@
                                                 <span class="comment-user">{{ comment.username }}</span>
                                                 <span class="comment-time">{{ comment.timestamp }}</span>
                                                 {% if comment.username == session.get('usuario_logado') %}
-                                                <form action="{{ url_for('delete_comment') }}" method="POST" class="delete-comment-form" onsubmit="return confirm('Apagar este comentário?')">
+                                                <form action="{{ url_for('main.delete_comment') }}" method="POST" class="delete-comment-form" onsubmit="return confirm('Apagar este comentário?')">
                                                     <input type="hidden" name="post_title" value="{{ post.title }}">
                                                     <input type="hidden" name="username" value="{{ comment.username }}">
                                                     <input type="hidden" name="timestamp" value="{{ comment.timestamp }}">
@@ -114,7 +114,7 @@
 
                         <!-- formulario do comentarios -->
                         <div class="comment-form-wrapper" style="display: block;">
-                            <form class="comment-form" action="{{ url_for('add_comment') }}" method="POST">
+                            <form class="comment-form" action="{{ url_for('main.add_comment') }}" method="POST">
                                 <input type="hidden" name="post_title" value="{{ post.title }}">
                                 <input type="hidden" name="redirect_to" value="{{ request.url }}">
                                 <div class="d-flex gap-2">
@@ -138,7 +138,7 @@
                 <div class="empty text-center py-5">
                     <i class="bi bi-inbox display-1 text-muted mb-3"></i>
                     <h3 class="fw-bold">Post não encontrado</h3>
-                    <a href="{{ url_for('home') }}" class="btn btn-primary btn-lg">
+                    <a href="{{ url_for('main.home') }}" class="btn btn-primary btn-lg">
                         Voltar ao Feed
                     </a>
                 </div>
