+++ templates/main/profile.html
@@ -58,7 +58,7 @@
                         <i class="bi bi-check"></i>
                         <span>Salvar Alterações</span>
                     </button>
-                    <a href="{{ url_for('home') }}" class="btn btn-outline-secondary profile-edit-cancel">
+                    <a href="{{ url_for('main.home') }}" class="btn btn-outline-secondary profile-edit-cancel">
                         <i class="bi bi-x"></i>
                         <span>Cancelar</span>
                     </a>
