+++ templates/main/publish.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends "base.html" %}
 
 {% block title %}Criar Publicação — InStruct{% endblock %}
 
@@ -17,12 +17,12 @@
         <aside class="col-12 col-md-3 col-lg-2 aside-area px-2 border-bottom border-md-0">
             <ul class="nav flex-column p-3">
                 <li class="nav-item mb-2">
-                    <a class="nav-link" href="{{ url_for('home') }}">
+                    <a class="nav-link" href="{{ url_for('main.home') }}">
                         <i class="bi bi-house"></i> Home
                     </a>
                 </li>
                 <li class="nav-item mb-2">
-                    <a class="nav-link active" href="{{ url_for('publish') }}">
+                    <a class="nav-link active" href="{{ url_for('main.publish') }}">
                         <i class="bi bi-plus-circle"></i> Criar Post
                     </a>
                 </li>
@@ -86,7 +86,7 @@
             </div>
             
             <!-- Forms -->
-            <form action="{{ url_for('send_publish_data') }}" method="POST">
+            <form action="{{ url_for('main.send_publish_data') }}" method="POST">
             <div class="tab-content">
                 <div class="tab-pane fade show active" id="description-tab-pane" role="tabpanel" aria-labelledby="description-tab" tabindex="0">
